Move Gitter API methods into a common module
